+++ avalanche/benchmarks/__init__.py
@@ -9,5 +9,5 @@ are made available.
 """
 
 from .scenarios import *
-from .generators import *
+from avalanche.benchmarks.scenarios.deprecated.generators import *
 from .classic import *
