--- avalanche/benchmarks/generators/__init__.py
@@ -1,1 +1,0 @@
-from .benchmark_generators import *
