+++ avalanche/benchmarks/scenarios/deprecated/classification_scenario.py
