+++ avalanche/benchmarks/scenarios/deprecated/lazy_dataset_sequence.py
