+++ avalanche/benchmarks/scenarios/deprecated/new_classes/__init__.py
