+++ avalanche/benchmarks/scenarios/deprecated/new_instances/__init__.py
