+++ avalanche/benchmarks/scenarios/deprecated/new_instances/ni_scenario.py
@@ -13,12 +13,12 @@ from typing import Optional, List, Sequence, Dict, Any
 
 import torch
 
-from avalanche.benchmarks.scenarios.classification_scenario import (
+from avalanche.benchmarks.scenarios.deprecated.classification_scenario import (
     ClassificationScenario,
     ClassificationStream,
     ClassificationExperience,
 )
-from avalanche.benchmarks.scenarios.new_instances.ni_utils import (
+from avalanche.benchmarks.scenarios.deprecated.new_instances.ni_utils import (
     _exp_structure_from_assignment,
 )
 from avalanche.benchmarks.utils import classification_subset
