+++ avalanche/benchmarks/scenarios/deprecated/new_instances/ni_utils.py
