+++ tests/benchmarks/scenarios/test_scenarios_typechecks.py
@@ -6,8 +6,8 @@ from os.path import expanduser
 from torchvision.datasets import MNIST
 
 from avalanche.benchmarks import tensors_benchmark
-from avalanche.benchmarks.generators import nc_benchmark, ni_benchmark
-from avalanche.benchmarks.scenarios.classification_scenario import (
+from avalanche.benchmarks.scenarios.deprecated.generators import nc_benchmark, ni_benchmark
+from avalanche.benchmarks.scenarios.deprecated.classification_scenario import (
     ClassificationExperience,
 )
