+++ tests/test_nc_mt_scenario.py
@@ -6,9 +6,9 @@ import torch
 from torchvision.datasets import MNIST
 
 from avalanche.benchmarks.datasets import default_dataset_location
-from avalanche.benchmarks.scenarios.new_classes import NCExperience
+from avalanche.benchmarks.scenarios.deprecated.new_classes import NCExperience
 from avalanche.benchmarks.utils import classification_subset
-from avalanche.benchmarks.scenarios.new_classes.nc_utils import (
+from avalanche.benchmarks.scenarios.deprecated.new_classes.nc_utils import (
     make_nc_transformation_subset,
 )
 from avalanche.benchmarks import nc_benchmark, ClassificationStream
